+++ udata_hydra/analysis/csv.py
@@ -99,7 +99,8 @@ async def notify_udata(check_id: int, table_name: str) -> None:
             },
         }
         if config.CSV_TO_PARQUET:
-            payload["parquet_id"] = table_name
+            payload["document"]["analysis:parsing:parquet_url"] = check.get("parquet_url")
+            payload["document"]["analysis:parsing:parquet_size"] = check.get("parquet_size")
         queue.enqueue(send, _priority="high", **payload)
 
 
@@ -163,20 +164,21 @@ async def analyse_csv(
         )
         timer.mark("csv-to-db")
 
-        await csv_to_parquet(
+        parquet_args: tuple[str, int] | None = await csv_to_parquet(
             file_path=tmp_file.name,
             inspection=csv_inspection,
             table_name=table_name,
             resource_id=resource_id,
         )
         timer.mark("csv-to-parquet")
-
         if check_id:
             await Check.update(
                 check_id,
                 {
                     "parsing_table": table_name,
                     "parsing_finished_at": datetime.now(timezone.utc),
+                    "parquet_url": parquet_args[0] if parquet_args else None,
+                    "parquet_size": parquet_args[1] if parquet_args else None,
                 },
             )
         await csv_to_db_index(table_name, csv_inspection, check)
@@ -269,13 +271,18 @@ async def csv_to_parquet(
     inspection: dict,
     table_name: str,
     resource_id: str | None = None,
-) -> None:
+) -> tuple[str, int] | None:
     """
     Convert a csv file to parquet using inspection data.
 
-    :file_path: CSV file path to convert
-    :inspection: CSV detective report
-    :table_name: used to name the parquet file
+    Args:
+        :file_path: CSV file path to convert
+        :inspection: CSV detective report
+        :table_name: used to name the parquet file
+
+    Returns:
+        :parquet_url: URL of the parquet file
+        :parquet_size: size of the parquet file
     """
     if not config.CSV_TO_PARQUET:
         log.debug("CSV_TO_PARQUET turned off, skipping parquet export.")
@@ -295,9 +302,11 @@ async def csv_to_parquet(
     parquet_file, _ = save_as_parquet(
         records=generate_records(file_path, inspection, columns),
         columns=columns,
-        output_name=table_name,
+        output_filename=table_name,
     )
-    minio_client.send_file(parquet_file)
+    parquet_size: int = os.path.getsize(parquet_file)
+    parquet_url: str = minio_client.send_file(parquet_file)
+    return parquet_url, parquet_size
 
 
 async def csv_to_db(